Fixed Codacy issues

+++ EmbyStat.Common/Extensions/RoundNumberExtension.cs
@@ -17,7 +17,7 @@ namespace EmbyStat.Common.Extensions
 
         public static int? RoundToFive(this double date)
         {
-            return (int)Math.Floor((double)date *20) * 5;
+            return (int)Math.Floor(date * 20) * 5;
         }
 
         public static double? RoundToHalf(this float? rating)
